feat: layout controller should support  &  option
