+++ packages/g6/package.json
@@ -63,7 +63,7 @@
     "@antv/g-svg": "^1.8.36",
     "@antv/g-webgl": "^1.7.44",
     "@antv/graphlib": "^2.0.0",
-    "@antv/layout": "^1.0.0-alpha.15",
+    "@antv/layout": "^1.0.0-alpha.17",
     "@antv/layout-gpu": "^1.0.0-alpha.3",
     "@antv/util": "~2.0.5",
     "typedoc-plugin-markdown": "^3.14.0"
