+++ packages/g6/src/types/graph.ts
@@ -6,10 +6,10 @@ import { AnimateCfg } from './animate';
 import { BehaviorObjectOptionsOf, BehaviorOptionsOf, BehaviorRegistry } from './behavior';
 import { ComboModel, ComboUserModel } from './combo';
 import { Padding, Point } from './common';
-import { DataChangeType, GraphData } from './data';
+import { GraphData } from './data';
 import { EdgeModel, EdgeUserModel } from './edge';
 import { ITEM_TYPE } from './item';
-import { LayoutCommonConfig } from './layout';
+import { LayoutOptions } from './layout';
 import { NodeModel, NodeUserModel } from './node';
 import { Specification } from './spec';
 import { FitViewRules, GraphAlignment } from './view';
@@ -91,7 +91,7 @@ export interface IGraph<B extends BehaviorRegistry = BehaviorRegistry> extends E
    * @returns
    * @group Data
    */
-  changeData: (data: GraphData, type: 'replace' | 'mergeReplace') => void;
+  changeData: (data: GraphData, type?: 'replace' | 'mergeReplace') => void;
   /**
    * Clear the graph, means remove all the items on the graph.
    * @returns
@@ -283,12 +283,8 @@ export interface IGraph<B extends BehaviorRegistry = BehaviorRegistry> extends E
    * @param {boolean} stack push it into stack
    * @group Layout
    */
-  layout: (
-    cfg?: LayoutCommonConfig,
-    align?: GraphAlignment,
-    canvasPoint?: Point,
-    stack?: boolean,
-  ) => void;
+  layout: (options?: LayoutOptions) => Promise<void>;
+  stopLayout: () => void;
 
   // ===== interaction =====
   /**
