+++ packages/g6/tests/datasets/dataset1.ts
@@ -1,0 +1,841 @@
+import { GraphData } from '../../src';
+
+const data: GraphData = {
+  nodes: [
+    {
+      id: 'Argentina',
+      data: {
+        name: 'Argentina',
+      },
+    },
+    {
+      id: 'Australia',
+      data: {
+        name: 'Australia',
+      },
+    },
+    {
+      id: 'Belgium',
+      data: {
+        name: 'Belgium',
+      },
+    },
+    {
+      id: 'Brazil',
+      data: {
+        name: 'Brazil',
+      },
+    },
+    {
+      id: 'Colombia',
+      data: {
+        name: 'Colombia',
+      },
+    },
+    {
+      id: 'Costa Rica',
+      data: {
+        name: 'Costa Rica',
+      },
+    },
+    {
+      id: 'Croatia',
+      data: {
+        name: 'Croatia',
+      },
+    },
+    {
+      id: 'Denmark',
+      data: {
+        name: 'Denmark',
+      },
+    },
+    {
+      id: 'Egypt',
+      data: {
+        name: 'Egypt',
+      },
+    },
+    {
+      id: 'England',
+      data: {
+        name: 'England',
+      },
+    },
+    {
+      id: 'France',
+      data: {
+        name: 'France',
+      },
+    },
+    {
+      id: 'Germany',
+      data: {
+        name: 'Germany',
+      },
+    },
+    {
+      id: 'Iceland',
+      data: {
+        name: 'Iceland',
+      },
+    },
+    {
+      id: 'IR Iran',
+      data: {
+        name: 'IR Iran',
+      },
+    },
+    {
+      id: 'Japan',
+      data: {
+        name: 'Japan',
+      },
+    },
+    {
+      id: 'Korea Republic',
+      data: {
+        name: 'Korea Republic',
+      },
+    },
+    {
+      id: 'Mexico',
+      data: {
+        name: 'Mexico',
+      },
+    },
+    {
+      id: 'Morocco',
+      data: {
+        name: 'Morocco',
+      },
+    },
+    {
+      id: 'Nigeria',
+      data: {
+        name: 'Nigeria',
+      },
+    },
+    {
+      id: 'Panama',
+      data: {
+        name: 'Panama',
+      },
+    },
+    {
+      id: 'Peru',
+      data: {
+        name: 'Peru',
+      },
+    },
+    {
+      id: 'Poland',
+      data: {
+        name: 'Poland',
+      },
+    },
+    {
+      id: 'Portugal',
+      data: {
+        name: 'Portugal',
+      },
+    },
+    {
+      id: 'Russia',
+      data: {
+        name: 'Russia',
+      },
+    },
+    {
+      id: 'Saudi Arabia',
+      data: {
+        name: 'Saudi Arabia',
+      },
+    },
+    {
+      id: 'Senegal',
+      data: {
+        name: 'Senegal',
+      },
+    },
+    {
+      id: 'Serbia',
+      data: {
+        name: 'Serbia',
+      },
+    },
+    {
+      id: 'Spain',
+      data: {
+        name: 'Spain',
+      },
+    },
+    {
+      id: 'Sweden',
+      data: {
+        name: 'Sweden',
+      },
+    },
+    {
+      id: 'Switzerland',
+      data: {
+        name: 'Switzerland',
+      },
+    },
+    {
+      id: 'Tunisia',
+      data: {
+        name: 'Tunisia',
+      },
+    },
+    {
+      id: 'Uruguay',
+      data: {
+        name: 'Uruguay',
+      },
+    },
+  ],
+  edges: [
+    {
+      id: '0',
+      target: 'Russia',
+      source: 'Saudi Arabia',
+      data: {
+        target_score: 5,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '1',
+      target: 'Uruguay',
+      source: 'Egypt',
+      data: {
+        target_score: 1,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '2',
+      target: 'Russia',
+      source: 'Egypt',
+      data: {
+        target_score: 3,
+        source_score: 1,
+        directed: true,
+      },
+    },
+    {
+      id: '3',
+      target: 'Uruguay',
+      source: 'Saudi Arabia',
+      data: {
+        target_score: 1,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '4',
+      target: 'Uruguay',
+      source: 'Russia',
+      data: {
+        target_score: 3,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '5',
+      target: 'Saudi Arabia',
+      source: 'Egypt',
+      data: {
+        target_score: 2,
+        source_score: 1,
+        directed: true,
+      },
+    },
+    {
+      id: '6',
+      target: 'IR Iran',
+      source: 'Morocco',
+      data: {
+        target_score: 1,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '7',
+      target: 'Portugal',
+      source: 'Spain',
+      data: {
+        target_score: 3,
+        source_score: 3,
+        directed: false,
+      },
+    },
+    {
+      id: '8',
+      target: 'Portugal',
+      source: 'Morocco',
+      data: {
+        target_score: 1,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '9',
+      target: 'Spain',
+      source: 'IR Iran',
+      data: {
+        target_score: 1,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '10',
+      target: 'IR Iran',
+      source: 'Portugal',
+      data: {
+        target_score: 1,
+        source_score: 1,
+        directed: false,
+      },
+    },
+    {
+      id: '11',
+      target: 'Spain',
+      source: 'Morocco',
+      data: {
+        target_score: 2,
+        source_score: 2,
+        directed: false,
+      },
+    },
+    {
+      id: '12',
+      target: 'France',
+      source: 'Australia',
+      data: {
+        target_score: 2,
+        source_score: 1,
+        directed: true,
+      },
+    },
+    {
+      id: '13',
+      target: 'Denmark',
+      source: 'Peru',
+      data: {
+        target_score: 1,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '14',
+      target: 'Denmark',
+      source: 'Australia',
+      data: {
+        target_score: 1,
+        source_score: 1,
+        directed: false,
+      },
+    },
+    {
+      id: '15',
+      target: 'France',
+      source: 'Peru',
+      data: {
+        target_score: 1,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '16',
+      target: 'Denmark',
+      source: 'France',
+      data: {
+        target_score: 0,
+        source_score: 0,
+        directed: false,
+      },
+    },
+    {
+      id: '17',
+      target: 'Peru',
+      source: 'Australia',
+      data: {
+        target_score: 2,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '18',
+      target: 'Argentina',
+      source: 'Iceland',
+      data: {
+        target_score: 1,
+        source_score: 1,
+      },
+    },
+    {
+      id: '19',
+      target: 'Croatia',
+      source: 'Nigeria',
+      data: {
+        target_score: 2,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '20',
+      target: 'Croatia',
+      source: 'Argentina',
+      data: {
+        target_score: 3,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '21',
+      target: 'Nigeria',
+      source: 'Iceland',
+      data: {
+        target_score: 2,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '22',
+      target: 'Argentina',
+      source: 'Nigeria',
+      data: {
+        target_score: 2,
+        source_score: 1,
+        directed: true,
+      },
+    },
+    {
+      id: '23',
+      target: 'Croatia',
+      source: 'Iceland',
+      data: {
+        target_score: 2,
+        source_score: 1,
+        directed: true,
+      },
+    },
+    {
+      id: '24',
+      target: 'Serbia',
+      source: 'Costa Rica',
+      data: {
+        target_score: 1,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '25',
+      target: 'Brazil',
+      source: 'Switzerland',
+      data: {
+        target_score: 1,
+        source_score: 1,
+        directed: false,
+      },
+    },
+    {
+      id: '26',
+      target: 'Brazil',
+      source: 'Costa Rica',
+      data: {
+        target_score: 2,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '27',
+      target: 'Switzerland',
+      source: 'Serbia',
+      data: {
+        target_score: 2,
+        source_score: 1,
+        directed: true,
+      },
+    },
+    {
+      id: '28',
+      target: 'Brazil',
+      source: 'Serbia',
+      data: {
+        target_score: 2,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '29',
+      target: 'Switzerland',
+      source: 'Costa Rica',
+      data: {
+        target_score: 2,
+        source_score: 2,
+        directed: false,
+      },
+    },
+    {
+      id: '30',
+      target: 'Mexico',
+      source: 'Germany',
+      data: {
+        target_score: 1,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '31',
+      target: 'Sweden',
+      source: 'Korea Republic',
+      data: {
+        target_score: 1,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '32',
+      target: 'Mexico',
+      source: 'Korea Republic',
+      data: {
+        target_score: 1,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '33',
+      target: 'Germany',
+      source: 'Sweden',
+      data: {
+        target_score: 2,
+        source_score: 1,
+        directed: true,
+      },
+    },
+    {
+      id: '34',
+      target: 'Korea Republic',
+      source: 'Germany',
+      data: {
+        target_score: 2,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '35',
+      target: 'Sweden',
+      source: 'Mexico',
+      data: {
+        target_score: 3,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '36',
+      target: 'Belgium',
+      source: 'Panama',
+      data: {
+        target_score: 3,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '37',
+      target: 'England',
+      source: 'Tunisia',
+      data: {
+        target_score: 2,
+        source_score: 1,
+        directed: true,
+      },
+    },
+    {
+      id: '38',
+      target: 'Belgium',
+      source: 'Tunisia',
+      data: {
+        target_score: 5,
+        source_score: 2,
+        directed: true,
+      },
+    },
+    {
+      id: '39',
+      target: 'England',
+      source: 'Panama',
+      data: {
+        target_score: 6,
+        source_score: 1,
+        directed: true,
+      },
+    },
+    {
+      id: '40',
+      target: 'Belgium',
+      source: 'England',
+      data: {
+        target_score: 1,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '41',
+      target: 'Tunisia',
+      source: 'Panama',
+      data: {
+        target_score: 2,
+        source_score: 1,
+        directed: true,
+      },
+    },
+    {
+      id: '42',
+      target: 'Japan',
+      source: 'Colombia',
+      data: {
+        target_score: 2,
+        source_score: 1,
+        directed: true,
+      },
+    },
+    {
+      id: '43',
+      target: 'Senegal',
+      source: 'Poland',
+      data: {
+        target_score: 2,
+        source_score: 1,
+        directed: true,
+      },
+    },
+    {
+      id: '44',
+      target: 'Japan',
+      source: 'Senegal',
+      data: {
+        target_score: 2,
+        source_score: 2,
+        directed: false,
+      },
+    },
+    {
+      id: '45',
+      target: 'Colombia',
+      source: 'Poland',
+      data: {
+        target_score: 3,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '46',
+      target: 'Poland',
+      source: 'Japan',
+      data: {
+        target_score: 1,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '47',
+      target: 'Colombia',
+      source: 'Senegal',
+      data: {
+        target_score: 1,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '48',
+      target: 'Uruguay',
+      source: 'Portugal',
+      data: {
+        target_score: 2,
+        source_score: 1,
+        directed: true,
+      },
+    },
+    {
+      id: '49',
+      target: 'France',
+      source: 'Argentina',
+      data: {
+        target_score: 4,
+        source_score: 3,
+        directed: true,
+      },
+    },
+    {
+      id: '50',
+      target: 'Russia',
+      source: 'Spain',
+      data: {
+        target_score: 5,
+        source_score: 4,
+        directed: true,
+      },
+    },
+    {
+      id: '51',
+      target: 'Croatia',
+      source: 'Denmark',
+      data: {
+        target_score: 4,
+        source_score: 3,
+        directed: true,
+      },
+    },
+    {
+      id: '52',
+      target: 'Brazil',
+      source: 'Mexico',
+      data: {
+        target_score: 2,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '53',
+      target: 'Belgium',
+      source: 'Japan',
+      data: {
+        target_score: 3,
+        source_score: 2,
+        directed: true,
+      },
+    },
+    {
+      id: '54',
+      target: 'Sweden',
+      source: 'Switzerland',
+      data: {
+        target_score: 1,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '55',
+      target: 'England',
+      source: 'Colombia',
+      data: {
+        target_score: 4,
+        source_score: 3,
+        directed: true,
+      },
+    },
+    {
+      id: '56',
+      target: 'France',
+      source: 'Uruguay',
+      data: {
+        target_score: 2,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '57',
+      target: 'Belgium',
+      source: 'Brazil',
+      data: {
+        target_score: 2,
+        source_score: 1,
+        directed: true,
+      },
+    },
+    {
+      id: '58',
+      target: 'Croatia',
+      source: 'Russia',
+      data: {
+        target_score: 6,
+        source_score: 5,
+        directed: true,
+      },
+    },
+    {
+      id: '59',
+      target: 'England',
+      source: 'Sweden',
+      data: {
+        target_score: 2,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '60',
+      target: 'France',
+      source: 'Belgium',
+      data: {
+        target_score: 1,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '61',
+      target: 'Croatia',
+      source: 'England',
+      data: {
+        target_score: 2,
+        source_score: 1,
+        directed: true,
+      },
+    },
+    {
+      id: '62',
+      target: 'Belgium',
+      source: 'England',
+      data: {
+        target_score: 2,
+        source_score: 0,
+        directed: true,
+      },
+    },
+    {
+      id: '63',
+      target: 'France',
+      source: 'Croatia',
+      data: {
+        target_score: 4,
+        source_score: 2,
+        directed: true,
+      },
+    },
+  ],
+};
+
+export { data };
