+++ packages/g6/tests/unit/layout-spec.ts
@@ -1,844 +1,7 @@
-import G6, { GraphData, IGraph } from '../../src/index';
+import G6, { IGraph } from '../../src/index';
+import { data } from '../datasets/dataset1';
 const container = document.createElement('div');
 document.querySelector('body').appendChild(container);
-
-const data: GraphData = {
-  nodes: [
-    {
-      id: 'Argentina',
-      data: {
-        name: 'Argentina',
-      },
-    },
-    {
-      id: 'Australia',
-      data: {
-        name: 'Australia',
-      },
-    },
-    {
-      id: 'Belgium',
-      data: {
-        name: 'Belgium',
-      },
-    },
-    {
-      id: 'Brazil',
-      data: {
-        name: 'Brazil',
-      },
-    },
-    {
-      id: 'Colombia',
-      data: {
-        name: 'Colombia',
-      },
-    },
-    {
-      id: 'Costa Rica',
-      data: {
-        name: 'Costa Rica',
-      },
-    },
-    {
-      id: 'Croatia',
-      data: {
-        name: 'Croatia',
-      },
-    },
-    {
-      id: 'Denmark',
-      data: {
-        name: 'Denmark',
-      },
-    },
-    {
-      id: 'Egypt',
-      data: {
-        name: 'Egypt',
-      },
-    },
-    {
-      id: 'England',
-      data: {
-        name: 'England',
-      },
-    },
-    {
-      id: 'France',
-      data: {
-        name: 'France',
-      },
-    },
-    {
-      id: 'Germany',
-      data: {
-        name: 'Germany',
-      },
-    },
-    {
-      id: 'Iceland',
-      data: {
-        name: 'Iceland',
-      },
-    },
-    {
-      id: 'IR Iran',
-      data: {
-        name: 'IR Iran',
-      },
-    },
-    {
-      id: 'Japan',
-      data: {
-        name: 'Japan',
-      },
-    },
-    {
-      id: 'Korea Republic',
-      data: {
-        name: 'Korea Republic',
-      },
-    },
-    {
-      id: 'Mexico',
-      data: {
-        name: 'Mexico',
-      },
-    },
-    {
-      id: 'Morocco',
-      data: {
-        name: 'Morocco',
-      },
-    },
-    {
-      id: 'Nigeria',
-      data: {
-        name: 'Nigeria',
-      },
-    },
-    {
-      id: 'Panama',
-      data: {
-        name: 'Panama',
-      },
-    },
-    {
-      id: 'Peru',
-      data: {
-        name: 'Peru',
-      },
-    },
-    {
-      id: 'Poland',
-      data: {
-        name: 'Poland',
-      },
-    },
-    {
-      id: 'Portugal',
-      data: {
-        name: 'Portugal',
-      },
-    },
-    {
-      id: 'Russia',
-      data: {
-        name: 'Russia',
-      },
-    },
-    {
-      id: 'Saudi Arabia',
-      data: {
-        name: 'Saudi Arabia',
-      },
-    },
-    {
-      id: 'Senegal',
-      data: {
-        name: 'Senegal',
-      },
-    },
-    {
-      id: 'Serbia',
-      data: {
-        name: 'Serbia',
-      },
-    },
-    {
-      id: 'Spain',
-      data: {
-        name: 'Spain',
-      },
-    },
-    {
-      id: 'Sweden',
-      data: {
-        name: 'Sweden',
-      },
-    },
-    {
-      id: 'Switzerland',
-      data: {
-        name: 'Switzerland',
-      },
-    },
-    {
-      id: 'Tunisia',
-      data: {
-        name: 'Tunisia',
-      },
-    },
-    {
-      id: 'Uruguay',
-      data: {
-        name: 'Uruguay',
-      },
-    },
-  ],
-  edges: [
-    {
-      id: '0',
-      target: 'Russia',
-      source: 'Saudi Arabia',
-      data: {
-        target_score: 5,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '1',
-      target: 'Uruguay',
-      source: 'Egypt',
-      data: {
-        target_score: 1,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '2',
-      target: 'Russia',
-      source: 'Egypt',
-      data: {
-        target_score: 3,
-        source_score: 1,
-        directed: true,
-      },
-    },
-    {
-      id: '3',
-      target: 'Uruguay',
-      source: 'Saudi Arabia',
-      data: {
-        target_score: 1,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '4',
-      target: 'Uruguay',
-      source: 'Russia',
-      data: {
-        target_score: 3,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '5',
-      target: 'Saudi Arabia',
-      source: 'Egypt',
-      data: {
-        target_score: 2,
-        source_score: 1,
-        directed: true,
-      },
-    },
-    {
-      id: '6',
-      target: 'IR Iran',
-      source: 'Morocco',
-      data: {
-        target_score: 1,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '7',
-      target: 'Portugal',
-      source: 'Spain',
-      data: {
-        target_score: 3,
-        source_score: 3,
-        directed: false,
-      },
-    },
-    {
-      id: '8',
-      target: 'Portugal',
-      source: 'Morocco',
-      data: {
-        target_score: 1,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '9',
-      target: 'Spain',
-      source: 'IR Iran',
-      data: {
-        target_score: 1,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '10',
-      target: 'IR Iran',
-      source: 'Portugal',
-      data: {
-        target_score: 1,
-        source_score: 1,
-        directed: false,
-      },
-    },
-    {
-      id: '11',
-      target: 'Spain',
-      source: 'Morocco',
-      data: {
-        target_score: 2,
-        source_score: 2,
-        directed: false,
-      },
-    },
-    {
-      id: '12',
-      target: 'France',
-      source: 'Australia',
-      data: {
-        target_score: 2,
-        source_score: 1,
-        directed: true,
-      },
-    },
-    {
-      id: '13',
-      target: 'Denmark',
-      source: 'Peru',
-      data: {
-        target_score: 1,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '14',
-      target: 'Denmark',
-      source: 'Australia',
-      data: {
-        target_score: 1,
-        source_score: 1,
-        directed: false,
-      },
-    },
-    {
-      id: '15',
-      target: 'France',
-      source: 'Peru',
-      data: {
-        target_score: 1,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '16',
-      target: 'Denmark',
-      source: 'France',
-      data: {
-        target_score: 0,
-        source_score: 0,
-        directed: false,
-      },
-    },
-    {
-      id: '17',
-      target: 'Peru',
-      source: 'Australia',
-      data: {
-        target_score: 2,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '18',
-      target: 'Argentina',
-      source: 'Iceland',
-      data: {
-        target_score: 1,
-        source_score: 1,
-      },
-    },
-    {
-      id: '19',
-      target: 'Croatia',
-      source: 'Nigeria',
-      data: {
-        target_score: 2,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '20',
-      target: 'Croatia',
-      source: 'Argentina',
-      data: {
-        target_score: 3,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '21',
-      target: 'Nigeria',
-      source: 'Iceland',
-      data: {
-        target_score: 2,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '22',
-      target: 'Argentina',
-      source: 'Nigeria',
-      data: {
-        target_score: 2,
-        source_score: 1,
-        directed: true,
-      },
-    },
-    {
-      id: '23',
-      target: 'Croatia',
-      source: 'Iceland',
-      data: {
-        target_score: 2,
-        source_score: 1,
-        directed: true,
-      },
-    },
-    {
-      id: '24',
-      target: 'Serbia',
-      source: 'Costa Rica',
-      data: {
-        target_score: 1,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '25',
-      target: 'Brazil',
-      source: 'Switzerland',
-      data: {
-        target_score: 1,
-        source_score: 1,
-        directed: false,
-      },
-    },
-    {
-      id: '26',
-      target: 'Brazil',
-      source: 'Costa Rica',
-      data: {
-        target_score: 2,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '27',
-      target: 'Switzerland',
-      source: 'Serbia',
-      data: {
-        target_score: 2,
-        source_score: 1,
-        directed: true,
-      },
-    },
-    {
-      id: '28',
-      target: 'Brazil',
-      source: 'Serbia',
-      data: {
-        target_score: 2,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '29',
-      target: 'Switzerland',
-      source: 'Costa Rica',
-      data: {
-        target_score: 2,
-        source_score: 2,
-        directed: false,
-      },
-    },
-    {
-      id: '30',
-      target: 'Mexico',
-      source: 'Germany',
-      data: {
-        target_score: 1,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '31',
-      target: 'Sweden',
-      source: 'Korea Republic',
-      data: {
-        target_score: 1,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '32',
-      target: 'Mexico',
-      source: 'Korea Republic',
-      data: {
-        target_score: 1,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '33',
-      target: 'Germany',
-      source: 'Sweden',
-      data: {
-        target_score: 2,
-        source_score: 1,
-        directed: true,
-      },
-    },
-    {
-      id: '34',
-      target: 'Korea Republic',
-      source: 'Germany',
-      data: {
-        target_score: 2,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '35',
-      target: 'Sweden',
-      source: 'Mexico',
-      data: {
-        target_score: 3,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '36',
-      target: 'Belgium',
-      source: 'Panama',
-      data: {
-        target_score: 3,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '37',
-      target: 'England',
-      source: 'Tunisia',
-      data: {
-        target_score: 2,
-        source_score: 1,
-        directed: true,
-      },
-    },
-    {
-      id: '38',
-      target: 'Belgium',
-      source: 'Tunisia',
-      data: {
-        target_score: 5,
-        source_score: 2,
-        directed: true,
-      },
-    },
-    {
-      id: '39',
-      target: 'England',
-      source: 'Panama',
-      data: {
-        target_score: 6,
-        source_score: 1,
-        directed: true,
-      },
-    },
-    {
-      id: '40',
-      target: 'Belgium',
-      source: 'England',
-      data: {
-        target_score: 1,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '41',
-      target: 'Tunisia',
-      source: 'Panama',
-      data: {
-        target_score: 2,
-        source_score: 1,
-        directed: true,
-      },
-    },
-    {
-      id: '42',
-      target: 'Japan',
-      source: 'Colombia',
-      data: {
-        target_score: 2,
-        source_score: 1,
-        directed: true,
-      },
-    },
-    {
-      id: '43',
-      target: 'Senegal',
-      source: 'Poland',
-      data: {
-        target_score: 2,
-        source_score: 1,
-        directed: true,
-      },
-    },
-    {
-      id: '44',
-      target: 'Japan',
-      source: 'Senegal',
-      data: {
-        target_score: 2,
-        source_score: 2,
-        directed: false,
-      },
-    },
-    {
-      id: '45',
-      target: 'Colombia',
-      source: 'Poland',
-      data: {
-        target_score: 3,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '46',
-      target: 'Poland',
-      source: 'Japan',
-      data: {
-        target_score: 1,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '47',
-      target: 'Colombia',
-      source: 'Senegal',
-      data: {
-        target_score: 1,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '48',
-      target: 'Uruguay',
-      source: 'Portugal',
-      data: {
-        target_score: 2,
-        source_score: 1,
-        directed: true,
-      },
-    },
-    {
-      id: '49',
-      target: 'France',
-      source: 'Argentina',
-      data: {
-        target_score: 4,
-        source_score: 3,
-        directed: true,
-      },
-    },
-    {
-      id: '50',
-      target: 'Russia',
-      source: 'Spain',
-      data: {
-        target_score: 5,
-        source_score: 4,
-        directed: true,
-      },
-    },
-    {
-      id: '51',
-      target: 'Croatia',
-      source: 'Denmark',
-      data: {
-        target_score: 4,
-        source_score: 3,
-        directed: true,
-      },
-    },
-    {
-      id: '52',
-      target: 'Brazil',
-      source: 'Mexico',
-      data: {
-        target_score: 2,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '53',
-      target: 'Belgium',
-      source: 'Japan',
-      data: {
-        target_score: 3,
-        source_score: 2,
-        directed: true,
-      },
-    },
-    {
-      id: '54',
-      target: 'Sweden',
-      source: 'Switzerland',
-      data: {
-        target_score: 1,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '55',
-      target: 'England',
-      source: 'Colombia',
-      data: {
-        target_score: 4,
-        source_score: 3,
-        directed: true,
-      },
-    },
-    {
-      id: '56',
-      target: 'France',
-      source: 'Uruguay',
-      data: {
-        target_score: 2,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '57',
-      target: 'Belgium',
-      source: 'Brazil',
-      data: {
-        target_score: 2,
-        source_score: 1,
-        directed: true,
-      },
-    },
-    {
-      id: '58',
-      target: 'Croatia',
-      source: 'Russia',
-      data: {
-        target_score: 6,
-        source_score: 5,
-        directed: true,
-      },
-    },
-    {
-      id: '59',
-      target: 'England',
-      source: 'Sweden',
-      data: {
-        target_score: 2,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '60',
-      target: 'France',
-      source: 'Belgium',
-      data: {
-        target_score: 1,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '61',
-      target: 'Croatia',
-      source: 'England',
-      data: {
-        target_score: 2,
-        source_score: 1,
-        directed: true,
-      },
-    },
-    {
-      id: '62',
-      target: 'Belgium',
-      source: 'England',
-      data: {
-        target_score: 2,
-        source_score: 0,
-        directed: true,
-      },
-    },
-    {
-      id: '63',
-      target: 'France',
-      source: 'Croatia',
-      data: {
-        target_score: 4,
-        source_score: 2,
-        directed: true,
-      },
-    },
-  ],
-};
 
 describe('layout', () => {
   let graph: IGraph<any>;
@@ -855,8 +18,205 @@ describe('layout', () => {
         radius: 200,
       },
     });
-    graph.on('afterrender', () => {
-      // expect(graph.canvas.document.childNodes[0].childNodes.length).toBe(2);
+
+    graph.once('afterlayout', () => {
+      const nodesData = graph.getAllNodesData();
+      expect(nodesData[0].data.x).toBe(450);
+      expect(nodesData[0].data.y).toBe(250);
+
+      graph.destroy();
+      done();
+    });
+  });
+
+  it('should trigger re-layout by calling `layout` method manually.', (done) => {
+    graph = new G6.Graph({
+      container,
+      width: 500,
+      height: 500,
+      type: 'graph',
+      data,
+      layout: {
+        type: 'circular',
+        center: [250, 250],
+        radius: 200,
+      },
+    });
+
+    // first-time layout
+    graph.once('afterlayout', () => {
+      const nodesData = graph.getAllNodesData();
+      expect(nodesData[0].data.x).toBe(450);
+      expect(nodesData[0].data.y).toBe(250);
+
+      // re-layout
+      graph.once('afterlayout', () => {
+        const nodesData = graph.getAllNodesData();
+        expect(nodesData[0].data.x).toBe(350);
+        expect(nodesData[0].data.y).toBe(250);
+
+        graph.destroy();
+        done();
+      });
+
+      graph.layout({
+        type: 'circular',
+        center: [250, 250],
+        radius: 100, // change radius here
+      });
+    });
+  });
+
+  it('should trigger re-layout by calling `changeData` method manually.', (done) => {
+    graph = new G6.Graph({
+      container,
+      width: 500,
+      height: 500,
+      type: 'graph',
+      data,
+      layout: {
+        type: 'circular',
+        center: [250, 250],
+        radius: 200,
+      },
+    });
+
+    // first-time layout
+    graph.once('afterlayout', () => {
+      const nodesData = graph.getAllNodesData();
+      expect(nodesData[0].data.x).toBe(450);
+      expect(nodesData[0].data.y).toBe(250);
+
+      // re-layout
+      graph.once('afterlayout', () => {
+        const nodesData = graph.getAllNodesData();
+        expect(nodesData[0].data.x).toBe(250);
+        expect(nodesData[0].data.y).toBe(250);
+
+        graph.destroy();
+        done();
+      });
+
+      // Only one single node.
+      const newData = {
+        nodes: [{ id: 'node13', data: { x: 50, y: 50 } }],
+        edges: [{ id: 'edge1', source: 'node13', target: 'node13', data: {} }],
+      };
+      graph.changeData(newData);
+    });
+  });
+
+  it('should run layout in WebWorker with `workerEnabled`.', (done) => {
+    graph = new G6.Graph({
+      container,
+      width: 500,
+      height: 500,
+      type: 'graph',
+      data,
+      layout: {
+        type: 'circular',
+        workerEnabled: true,
+        center: [250, 250],
+        radius: 200,
+      },
+    });
+
+    graph.once('afterlayout', () => {
+      const nodesData = graph.getAllNodesData();
+      expect(nodesData[0].data.x).toBe(450);
+      expect(nodesData[0].data.y).toBe(250);
+
+      // re-layout
+      graph.once('afterlayout', () => {
+        const nodesData = graph.getAllNodesData();
+        expect(nodesData[0].data.x).toBe(350);
+        expect(nodesData[0].data.y).toBe(250);
+
+        graph.destroy();
+        done();
+      });
+
+      graph.layout({
+        type: 'circular',
+        center: [250, 250],
+        radius: 100, // change radius here
+      });
+    });
+  });
+
+  it('should display the layout process with `animated`.', (done) => {
+    graph = new G6.Graph({
+      container,
+      width: 500,
+      height: 500,
+      type: 'graph',
+      data,
+      layout: {
+        type: 'd3force',
+        animated: true,
+        center: [250, 250],
+        preventOverlap: true,
+        nodeSize: 20,
+      },
+    });
+
+    graph.once('afterlayout', () => {
+      const nodesData = graph.getAllNodesData();
+      expect(nodesData.every((node) => node.data.x > 0 && node.data.y > 0)).toBeTruthy();
+      graph.destroy();
+      done();
+    });
+  });
+
+  it('should stop animated layout process with `stopLayout`.', (done) => {
+    graph = new G6.Graph({
+      container,
+      width: 500,
+      height: 500,
+      type: 'graph',
+      data,
+      layout: {
+        type: 'd3force',
+        animated: true,
+        center: [250, 250],
+        preventOverlap: true,
+        nodeSize: 20,
+      },
+    });
+
+    setTimeout(() => {
+      graph.stopLayout();
+
+      const nodesData = graph.getAllNodesData();
+      expect(nodesData.every((node) => node.data.x > 0 && node.data.y > 0)).toBeTruthy();
+
+      graph.destroy();
+      done();
+    }, 1000);
+  });
+
+  it('should manually steps the simulation with `iterations` and `animated` disabled.', (done) => {
+    graph = new G6.Graph({
+      container,
+      width: 500,
+      height: 500,
+      type: 'graph',
+      data,
+      layout: {
+        type: 'd3force',
+        animated: false,
+        center: [250, 250],
+        preventOverlap: true,
+        nodeSize: 20,
+        iterations: 1000,
+      },
+    });
+
+    graph.once('afterlayout', () => {
+      const nodesData = graph.getAllNodesData();
+      expect(nodesData.every((node) => node.data.x > 0 && node.data.y > 0)).toBeTruthy();
+
+      graph.destroy();
       done();
     });
   });
